Add support for i8's.
https://github.com/vellvm/vellvm/issues/24

This is implemented in the same way as i64s. Ideally we will improve
this in the future to work with arbitrary bit widths, since this is
something that LLVM supports, and also there is a lot of code
duplication currently.

One potential issue is that conversions involve a fair amount of
finnicky copy and pasting, and it would be nice to have this all
automated.