Starting to get the TopLevel ready for extraction.
Lots of busy work still needs to get done. Slowly starting to figure
out extraction, though.

Refactored the modules a little bit. Not sure they make a lot of
sense... I don't think we're going to extract a module for use, we are
just going to replace all functions with the ones in the appropriate
module in Cerberus, though perhaps it helps us organize things.

memD is sort of in a "working" state now. We need a definition of
mfix (do we need to thunkify this?), and runMemM, which will likely
be implemented with runND in Cerberus's smt2.ml file. It looks like
we can conceivably have a function like runMemM, though.

The interpreter.ml file has some dtyp to ctype0 conversion
stuff. Mostly playing around with it for extraction.