Getting rid of endianess in memory model...
Need a way to serialize structures that are the result of an
expression (e.g., select statements). To do this I am planning on just
having ExtractByte uv 0 represent the first byte stored in the memory,
not the least significant byte. That way I can just serialize
everything in memory as ExtractByte uv 1 ... ExtractByte uv
sz.

This does mean that endianess will have to be  accounted for in
concretize,  however...