Messing with modules for PtoI / ItoP and provenance.
I think I need to be able to do these kinds of things in order to
define concretizations with UVALUE_ExtractBytes / UVALUE_ConcatBytes.

In particular, concretization of ConcatBytes in the non-trivial cases
involves actually extracting a byte value from each of the uvalue
bytes. It makes sense to do this for the integer that corresponds to a
pointer, so I need ptr_to_int for this.

Similarly, if I have a list of bytes that is being deserialized into a
ptr, I need to be able to use int_to_ptr. int_to_ptr may also involve
a provenance.