Changing intrinsics to uvalues for now...
This seems necessary for intrinsics like memcpy that should preserve
nondeterminism in the bytes that are copied...

That said, it is currently a little unclear how some of these
intrinsics should work. For instance, llvm_fabs_f32 takes a float and
produces the absolute value of the float. Currently this is left
incomplete, as the semantic_function for it will fail unless it gets a
concrete value (along with other intrinsics). These intrinsics should
probably result in a pick event for their arguments.