Add real finite memory for #334.
Previously all memory models could technically have infinite pointers,
and pointers could be arbitrarily large.

This set of changes makes sure we have separate finite and infinite
addresses (these can be found in FiniteAddresses.v and
InfiniteAddresses.v respectively).

Note: there's still some fiddly naming and oddities, like both the
infinite and finite memory models are instantiated with a "finite
sizeof" module, and are both built from "FiniteMemory" modules. This
is sort of just a quirk of naming, that maybe we should change, though
it's not quite as mad as it seems at first glance as the infinite
memory model is in some sense a finite memory model which happens to
have some extra properties, which we get by instantiating some
extra infinite memory model modules.

Changes:

- int_to_ptr can now OOM.
- handle_gep can now OOM as a result of the int_to_ptr changes.
- Some lemmas about OOM had to be added to the GepM module.
- Had to adjust LLVMParamsBig to take into account int_to_ptr casts
  always succeeding on infinite memory models.
- Make frames / heaps keep track of addresses, instead of integers.
  + This has made it possible to prove:
  + free_frame_memory_byte_disjoint_allocated
  +  free_block_memory_byte_disjoint_allocated
- Added int_ptr_ptr_to_int_exists in ITOP module... Didn't end up being
  necessary, but could be useful in the future.
- Split FiniteAddresses.v into FiniteAddresses.v and InfiniteAddresses.v
- Add InfToFin and FinToInf address conversions in InfiniteToFinite.v
- Fix up interpreter.ml to use infinite addresses.

The following lemmas had to be reproved because of the int_to_ptr and
handle_gep changes:

- get_consecutive_ptrs_cons.
- get_consecutive_ptrs_ge.
- get_consecutive_ptrs_range.
- get_consecutive_ptrs_nth.
- get_consecutive_ptrs_nth_eq1.
- get_consecutive_ptrs_inv.
- get_consecutive_ptrs_no_ub.
- exec_correct_get_consecutive_pointers.
- get_consecutive_ptrs_succeeds.
- proofs in FiniteMemoryModelSpecPrimitives.v
- heap_eqv_add
- MemMonad_run_get_consecutive_ptrs.
- find_free_block_correct.
- get_consecutive_ptrs_MemPropT_MemState
- get_consecutive_ptrs_MemPropT_eq1
- Proofs in FiniteExecPrimitives.v. Had troubles that resulted in me
  changing frames and heaps to store addresses instead of concrete Z values.