Convert ShowAST to DString.
- Add DList.v file for some difference list functions.
  + Trying to use as much tail recursion as possible
      in DLists to prevent stack performance problems.
- Add DString type (like string, but it uses difference lists)
- Convert show instances to use DString where possible.
  + Uses the new DShow typeclass and instances.

Additionally we added the QCShowTesting.v file

Compiling this file should cause a stack overflow with the old show
instances (that don't use diff lists for appends).

defSize can be changed to alter how large the programs we try to
generate are --- I am not sure if 50 is just too large or not. We may
have to compare the diff list version to the original version and see
what size tends to work for both of them. Ideally we should be able to
increase the size more with the diff list version :). Test cases are
randomly generated, so you may have to run the QuickChick command
multiple times to be sure (it may randomly decide to generate smaller programs).