Checkpoint for returning k_spec to its former glory.
InterpMemoryProp.v is now compiling. Had some difficulties with
finding the previous definition of k_spec... That version used a
k_spec_WF class, but it didn't make much sense to me... It seems like
it lost a sense of what k_spec_correct should be. This version
hopefully makes sense. It's possible I will need to mess with
k_spec_WF for future lemmas, but it worked to get through
InterpMemoryProp.v.