Fix bug with bitcast.
Concretizing bitcast would run into an issue where it would spit out
this message:

"Bitcast failed to convert to dvalue: Attempting to convert a
partially non-reduced uvalue to dvalue. Should not happen"

Obviously... This should not happen!

The bitcast was implemented by serializing to bytes and then
deserializing the bytes... Since this happens on dvalues I thought
that we should get a simple "uvalue that's actually just a dvalue"
that `uvalue_to_dvalue` would succeed with... Unfortunately this is
not the case because we construct a `UVALUE_ConcatBytes` when
deserializing bytes from one type to another. This change makes the
pure bitcasts return a uvalue instead which we then concretize in the
concretization functions.