More progress on correct IntMap memory model.
Issues are coming up with provenance and `find_free_block`... If we
allocate a free block and return pointers with a freshly allocated
provenance, then we likely have to return a new `Memory` as well,
because the fresh provenance value would modify memory...

The easiest solution would be to return Z values instead of addrs and
assume that our memory models would have integer based addresses, but
this would probably rule out logical memory models in our framework,
which isn't ideal.

Another solution is to start with a wildcard or nil provenance, and
fill in provenance later during actual allocations... This would keep
the memory the same in `find_free_block`.

A final option would be to return a new memory, which is potentially
modified for keeping track of provenance, and add lemmas mentioning
the preservation of reads and allocations in the new memory. This
might be the most "morally correct" option, but is probably the
biggest change.